Infrastructure | Implemented Workout Session Converter

+++ domain/src/main/kotlin/com/github/tomplum/activity/workout/Distance.kt
@@ -1,3 +1,3 @@
 package com.github.tomplum.activity.workout
 
-data class Distance(val unit: String, val value: Int)
+data class Distance(val unit: String, val value: Double)
